Refactor: Split async functions from sync wrappers
