+++ src/aleph_client/__init__.py
@@ -9,5 +9,3 @@ except DistributionNotFound:
     __version__ = "unknown"
 finally:
     del get_distribution, DistributionNotFound
-
-from .main import create_post, create_aggregate
