--- src/aleph_client/main.py
@@ -1,14 +1,0 @@
-from .asynchronous import (
-    sync_ipfs_push as ipfs_push,
-    sync_storage_push as storage_push,
-    sync_ipfs_push_file as ipfs_push_file,
-    sync_storage_push_file as storage_push_file,
-    sync_broadcast as broadcast,
-    sync_create_post as create_post,
-    sync_create_aggregate as create_aggregate,
-    sync_create_store as create_store,
-    sync_submit as submit,
-    sync_fetch_aggregate as fetch_aggregate,
-    sync_get_posts as get_posts,
-    sync_get_messages as get_messages,
-)
